Refactor: Split checker.rs into smaller modules
This split src/compiler/sema/checker.rs into mod.rs, decl.rs, stmt.rs, and expr.rs for better maintainability. Closes #15
